+++ CHANGES.rst
@@ -2,7 +2,20 @@ Changelog
 =========
 
 
-2.2.6 (unreleased)
+2.3.2 (unreleased)
+------------------
+
+- Add fullobjects in @bookings
+  [mamico]
+
+2.3.1 (2023-12-21)
+------------------
+
+- add content-transfer-encoding customization 
+  [mamico]
+
+
+2.3.0 (2023-12-19)
 ------------------
 
 - Add booking_type vocabularies for portal_root